Use compile to supply a filename
This allows us to not rely on the assumption that "<string>" is always
representing the filename of the main file. This led to problems in the
skipping logic where some code executed through an import statement was
assumed to be our code. That led to stack frames being popped (which
shouldn't be popped at that point).